AIMentalWellbeingChatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
pydantic==2.5.0
python-dotenv==1.0.0